% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeSVD.R
\name{computeSVD}
\alias{computeSVD}
\title{Title Perform SVD toward reduced matrix generated from a tensor with 
partial summation}
\usage{
computeSVD(matrix1, matrix2, dim = 10L, scale = TRUE)
}
\arguments{
\item{matrix1}{The first original matrix that generates a tensor}

\item{matrix2}{The second original matrix that generates a tensor}

\item{dim}{The number of singular value vectors to be computed}

\item{scale}{If matrix should be scaled or not}
}
\value{
Singular value vectors attributed to two sets of objects associated 
with singular value vectors attributed to features, by multiplying
}
\description{
Title Perform SVD toward reduced matrix generated from a tensor with 
partial summation
}
\examples{
matrix1 <- matrix(runif(200),20)
matrix2 <- matrix(runif(400),20)
SVD <- computeSVD(matrix1,matrix2)
}
