% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepareCondDrugandDisease.R
\name{prepareCondDrugandDisease}
\alias{prepareCondDrugandDisease}
\title{Prepare condition matrix for expDrug}
\usage{
prepareCondDrugandDisease(expDrug)
}
\arguments{
\item{expDrug}{input gene expression profile}
}
\value{
Condition matrix for expDrug
}
\description{
Prepare condition matrix for expDrug
}
\examples{
input <- readRDS(system.file(
  "extdata", "drug_disease_test_inputs.rds",
  package = "TDbasedUFEadv", mustWork = TRUE
))
Cancer_cell_lines <- input$Cancer_cell_lines[seq_len(3)]
Drug_and_Disease <- prepareexpDrugandDisease(Cancer_cell_lines)

Cond <- prepareCondDrugandDisease(Drug_and_Disease$expDrug)
}
