% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepareCondTCGA.R
\name{prepareCondTCGA}
\alias{prepareCondTCGA}
\title{Prepare Sample label for TCGA data}
\usage{
prepareCondTCGA(
  Multi_sample,
  Clinical,
  ID_column_of_Multi_sample,
  ID_column_of_Clinical
)
}
\arguments{
\item{Multi_sample}{list of sample ids}

\item{Clinical}{List of clinical data matrix from RTCGA.clinical}

\item{ID_column_of_Multi_sample}{Column numbers used for conditions}

\item{ID_column_of_Clinical}{Column numbers that include corresponding
sample ids in clinical data}
}
\value{
list of sample labels
}
\description{
Prepare Sample label for TCGA data
}
\examples{
example <- readRDS(system.file(
  "extdata", "enrichment_example.rds",
  package = "TDbasedUFEadv", mustWork = TRUE
))
cond <- prepareCondTCGA(
  example$Multi_sample, example$Clinical,
  rep(2, 4), rep(1, 4)
)
}
