% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrepareSummarizedExperimentTensorRect.R
\name{PrepareSummarizedExperimentTensorRect}
\alias{PrepareSummarizedExperimentTensorRect}
\title{Title Prepare tensor generated from two matricies that share samples}
\usage{
PrepareSummarizedExperimentTensorRect(
  sample,
  feature,
  value,
  featureRange = GRanges(NULL),
  sampleData = list(NULL)
)
}
\arguments{
\item{sample}{Chracter vecor of sample names}

\item{feature}{list of features from two matrices}

\item{value}{array, contents of}

\item{featureRange}{Genomic Ranges to be associated with features}

\item{sampleData}{List of conditional labeling associated with samples}
}
\value{
Tensor generated from two matricies that share samples
}
\description{
Title Prepare tensor generated from two matricies that share samples
}
\examples{
matrix1 <- matrix(runif(10000),200) #row features, column samples
matrix2 <- matrix(runif(20000),400) #row features, column samples
Z <- prepareTensorfromMatrix(t(matrix1),t(matrix2))
Z <- PrepareSummarizedExperimentTensorRect(sample=as.character(seq_len(50)),
feature=list(as.character(seq_len(200)),as.character(seq_len(400))),
sampleData=list(rep(seq_len(2),each=25)),value=Z)
}
