% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepareTensorfromMatrix.R
\name{prepareTensorfromMatrix}
\alias{prepareTensorfromMatrix}
\title{Title Generate tensor from two matricies}
\usage{
prepareTensorfromMatrix(matrix1, matrix2)
}
\arguments{
\item{matrix1}{the first input matrix}

\item{matrix2}{the second input matrix}
}
\value{
A tensor generated from the first and second matricies
}
\description{
Title Generate tensor from two matricies
}
\examples{
Z <- prepareTensorfromMatrix(matrix(runif(100),10),matrix(runif(100),10))
}
