% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectFeatureTransRect.R
\name{selectFeatureTransRect}
\alias{selectFeatureTransRect}
\title{Title Select features for a tensor generated from two matricies that
 share samples.}
\usage{
selectFeatureTransRect(
  HOSVD,
  cond,
  de = rep(1e-04, 2),
  p0 = 0.01,
  breaks = 100,
  input_all = NULL
)
}
\arguments{
\item{HOSVD}{HOSVD}

\item{cond}{list of conditions}

\item{de}{initial values for optimization of standard deviation}

\item{p0}{threshold value for the significance}

\item{breaks}{number of bins of the histgram of P-values}

\item{input_all}{The selected singular value vectors attributed to samples.
if NULL, interactive mode}
}
\value{
list of logical vector that represent if the individual features
are selected and P-values.
}
\description{
Title Select features for a tensor generated from two matricies that
 share samples.
}
\examples{
require(TDbasedUFE)
matrix1 <- matrix(runif(10000),200) #row features, column samples
matrix2 <- matrix(runif(20000),400) #row features, column samples
Z <- prepareTensorfromMatrix(t(matrix1),t(matrix2))
Z <- PrepareSummarizedExperimentTensorRect(sample=as.character(1:50),
feature=list(as.character(1:200),as.character(1:400)),
sampleData=list(rep(1:2,each=25)),value=Z)
HOSVD <- computeHosvd(Z)
cond <- list(attr(Z,"sampleData")[[1]],NULL,NULL)
index_all <- selectFeatureTransRect(HOSVD,cond,de=c(0.01,0.01),input_all=2)
}
