% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\docType{class}
\name{SummarizedExperimentTensorRect-class}
\alias{SummarizedExperimentTensorRect-class}
\title{Title}
\description{
Title
}
\section{Slots}{

\describe{
\item{\code{sample}}{character.}

\item{\code{feature}}{list.}

\item{\code{value}}{array.}

\item{\code{featureRange}}{GRanges.}

\item{\code{sampleData}}{list.}
}}

