% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TDbasedUFEadv-package.R
\docType{package}
\name{TDbasedUFEadv-package}
\alias{TDbasedUFEadv}
\alias{TDbasedUFEadv-package}
\title{TDbasedUFEadv: Advanced package of tensor decomposition based unsupervised feature extraction}
\description{
This is an advanced version of TDbasedUFE, which is a comprehensive package to perform Tensor decomposition based unsupervised feature extraction. In contrast to TDbasedUFE which can perform simple the feature selection and the multiomics analyses, this package can perform more complicated and advanced features, but they are not so popularly required. Only users who require more specific features can make use of its functionality.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/tagtag/TDbasedUFEadv}
  \item Report bugs at \url{https://github.com/tagtag/TDbasedUFEadv/issues}
}

}
\author{
\strong{Maintainer}: Y-h. Taguchi \email{tag@granular.com} (\href{https://orcid.org/0000-0003-0867-8986}{ORCID})

Authors:
\itemize{
  \item Y-h. Taguchi \email{tag@granular.com} (\href{https://orcid.org/0000-0003-0867-8986}{ORCID})
}

}
\keyword{internal}
