% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{TensorRect-class}
\alias{TensorRect-class}
\title{Class definitions}
\description{
Class definitions
}
\section{Slots}{

\describe{
\item{\code{sample}}{character.}

\item{\code{feature}}{list.}

\item{\code{value}}{array.}

\item{\code{featureRange}}{GRanges.}

\item{\code{sampleData}}{list.}
}}

