% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transSVD.R
\name{transSVD}
\alias{transSVD}
\title{Title Convert SVD to that for the case where samples are shared between two matrices}
\usage{
transSVD(SVD)
}
\arguments{
\item{SVD}{inout SVD object genetated from computeSVD function}
}
\value{
converted SVD objects
}
\description{
Title Convert SVD to that for the case where samples are shared between two matrices
}
\examples{
matrix1 <- matrix(runif(200),20)
matrix2 <- matrix(runif(400),20)
SVD <- computeSVD(matrix1,matrix2)
SVD <- transSVD(SVD)
}
