% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepareCondDrungandDisease.R
\name{prepareCondDrugandDisease}
\alias{prepareCondDrugandDisease}
\title{Title Prepare condition matrix for expDrug}
\usage{
prepareCondDrugandDisease(expDrug)
}
\arguments{
\item{expDrug}{input gene expression profile}
}
\value{
Condition matrix for expDrug
}
\description{
Title Prepare condition matrix for expDrug
}
\examples{
dummy <- prepareexpDrugandDisease()
expDrug <- dummy[[1]]
Cond <- prepareCondDrugandDisease(expDrug)
}
