% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepareCondDrungandDisease.R
\name{prepareCondDrugandDisease}
\alias{prepareCondDrugandDisease}
\title{Title Prepare condition matrix for expDrug}
\usage{
prepareCondDrugandDisease(expDrug)
}
\arguments{
\item{expDrug}{input gene expression profile}
}
\value{
Condition matrix for expDrug
}
\description{
Title Prepare condition matrix for expDrug
}
\examples{
\donttest{
require(RTCGA.rnaseq)
LIST <- list(ACC.rnaseq,
            BLCA.rnaseq,
            BRCA.rnaseq,
            CESC.rnaseq,
            COAD.rnaseq,
            ESCA.rnaseq,
            GBM.rnaseq,
            HNSC.rnaseq,
            KICH.rnaseq,
            KIRC.rnaseq,
            KIRP.rnaseq,
            LGG.rnaseq,
            LIHC.rnaseq,
            LUAD.rnaseq,
            LUSC.rnaseq,
            OV.rnaseq,
            PAAD.rnaseq,
            PCPG.rnaseq,
            PRAD.rnaseq,
            READ.rnaseq,
            SARC.rnaseq,
            SKCM.rnaseq,
            STAD.rnaseq,
            TGCT.rnaseq,
            THCA.rnaseq,
            UCEC.rnaseq,
            UCS.rnaseq)
dummy <- prepareexpDrugandDisease(LIST)
expDrug <- dummy[[1]]
Cond <- prepareCondDrugandDisease(expDrug)
}
}
