% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepareexpDrugandDisease.R
\name{prepareexpDrugandDisease}
\alias{prepareexpDrugandDisease}
\title{Title Generating gene expression of drug treated cell lines and a disease
cell line}
\usage{
prepareexpDrugandDisease(LIST)
}
\arguments{
\item{LIST}{list that includes indivisual data set from RTCGA.rnaseq}
}
\value{
list of expDrug and expDisease
}
\description{
Title Generating gene expression of drug treated cell lines and a disease
cell line
}
\examples{
 \donttest{
require(RTCGA.rnaseq)
LIST <- list(ACC.rnaseq,
            BLCA.rnaseq,
            BRCA.rnaseq)
dummy <- prepareexpDrugandDisease(LIST)
}
}
