% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepareexpDrugandDisease.R
\name{prepareexpDrugandDisease}
\alias{prepareexpDrugandDisease}
\title{Generating gene expression of drug treated cell lines and a disease
cell line}
\usage{
prepareexpDrugandDisease(Cancer_cell_lines)
}
\arguments{
\item{Cancer_cell_lines}{<- list(ACC.rnaseq,BLCA.rnaseq,BRCA.rnaseq)
list that includes individual data set from RTCGA.rnaseq}
}
\value{
list of expDrug and expDisease
}
\description{
Generating gene expression of drug treated cell lines and a disease
cell line
}
\examples{
 \donttest{
require(RTCGA.rnaseq)
Cancer_cell_lines <- list(ACC.rnaseq,BLCA.rnaseq,BRCA.rnaseq)
Drug_and_Disease <- prepareexpDrugandDisease(Cancer_cell_lines)
}
}
