% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepareexpDrugandDisease.R
\name{prepareexpDrugandDisease}
\alias{prepareexpDrugandDisease}
\title{Title Generating gene expression of drug treated cell lines and a disease
cell line}
\usage{
prepareexpDrugandDisease()
}
\value{
list of expDrug and expDisease
}
\description{
Title Generating gene expression of drug treated cell lines and a disease
cell line
}
\examples{
dummy <- prepareexpDrugandDisease()
}
