% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepareexpDrugandDisease.R
\name{prepareexpDrugandDisease}
\alias{prepareexpDrugandDisease}
\title{Title Generating gene expression of drug treated cell lines and a disease
cell line}
\usage{
prepareexpDrugandDisease(LIST)
}
\arguments{
\item{LIST}{list that includes indivisual data set from RTCGA.rnaseq}
}
\value{
list of expDrug and expDisease
}
\description{
Title Generating gene expression of drug treated cell lines and a disease
cell line
}
\examples{
 \donttest{
require(RTCGA.rnaseq)
LIST <- list(ACC.rnaseq,
            BLCA.rnaseq,
            BRCA.rnaseq,
            CESC.rnaseq,
            COAD.rnaseq,
            ESCA.rnaseq,
            GBM.rnaseq,
            HNSC.rnaseq,
            KICH.rnaseq,
            KIRC.rnaseq,
            KIRP.rnaseq,
            LGG.rnaseq,
            LIHC.rnaseq,
            LUAD.rnaseq,
            LUSC.rnaseq,
            OV.rnaseq,
            PAAD.rnaseq,
            PCPG.rnaseq,
            PRAD.rnaseq,
            READ.rnaseq,
            SARC.rnaseq,
            SKCM.rnaseq,
            STAD.rnaseq,
            TGCT.rnaseq,
            THCA.rnaseq,
            UCEC.rnaseq,
            UCS.rnaseq)
dummy <- prepareexpDrugandDisease(LIST)
}
}
