% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TDbasedUFEadv-package.R
\docType{package}
\name{TDbasedUFEadv-package}
\alias{TDbasedUFEadv}
\alias{TDbasedUFEadv-package}
\title{TDbasedUFEadv: TDbasedUFEadv}
\description{
More about what it does (maybe more than one line) Use four spaces when indenting paragraphs within the Description.
}
\keyword{internal}
