% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TDbasedUFEadv-package.R
\docType{package}
\name{TDbasedUFEadv-package}
\alias{TDbasedUFEadv}
\alias{TDbasedUFEadv-package}
\title{TDbasedUFEadv: TDbasedUFEadv}
\description{
More about what it does (maybe more than one line) This is a comprehensive package to perform Tensor decomposition based unsupervised feature extraction. It can perform unsupervised feature extraction. It uses tensor decompission. It is applicable to gene expression, DNA methylation, and histone modification etc. It can perform multiomics analysis. It is also applicable to single cell omics data sets.
}
\author{
\strong{Maintainer}: Taguchi Y-h. \email{tag@granular.com} (\href{https://orcid.org/0000-0003-0867-8986}{ORCID})

}
\keyword{internal}
